Restructure program entry points
I was going to try out optparse-applicative, so what I did here is a
stop-gap measure before that. It's nice to get rid of a second
executable, and also to embed the version number in the main exe in a
way that's easy for a user to see.

Cleaning up dependencies while I was at it.