Downgrade GHC, and add redundant calls to gather-and-publish
On Github Actions, macos-10.15 seems to not play well with the Stackage
snapshot with the very-new GHC 8.8: https://github.com/easoncxz/hack-assembler/runs/580182182

I'm going back to an older version of GHC to see whether this helps with
the build.

Meanwhile, I'm cleaning up the Travis config so that we get the new
Bottles for macos-10.13 and macos-10.14 right after Travis is done
building them, instead of waiting until Github Action finishes too.
Publish ASAP, I guess. It doesn't harm to run gather-and-publish
repeatedly, since it doesn't wipe out existing bottles if the sdist
version and sha256 are the same.